pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-segmentation"
version = "0.1.0"
description = "Pure PyTorch 2D vessel segmentation with VLM feature prior"
requires-python = ">=3.9"
# torch/torchvision must be installed separately:
#   uv pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
dependencies = [
    "numpy",
    "scipy",
    "scikit-image",
    "natsort",
    "opencv-python-headless",
    "pyyaml",
    "tqdm",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
